Feed intake monitor of pig performance test station

monitor_schedule() is deprecated. We will soon be totally

monitor_schedule(
  taskname,
  schedule,
  starttime,
  startdate,
  rscript_args = NULL,
  ...
)

Arguments

taskname

A character string with the name of the task. Defaults to the filename. Should not contain any spaces

schedule

Either one of 'ONCE', 'MONTHLY', 'WEEKLY', 'DAILY', 'HOURLY', 'MINUTE', 'ONLOGON', 'ONIDLE

starttime

A timepoint in HH:mm format indicating when to run the script. Defaults to within 62 seconds

startdate

A date that specifies the first date on which to run the task. Only applicable if schedule is of type 'MONTHLY', 'WEEKLY', 'DAILY', 'HOURLY', 'MINUTE'. Defaults to today in '%d/%m/%Y' format. Change to your locale format if needed

rscript_args

Character string with further arguments passed on to Rscript

...

other parameters

Value

CSV files in the path of save_path

Examples

monitor_schedule(
  taskname = "ppt_csv_monitor",
  schedule = "DAILY",
  starttime = "10:05",
  startdate = format(Sys.Date(), "%Y/%m/%d"),
  rscript_args = list(csv_path = "C:/Users/Dell/Downloads/test",
                      save_path = "C:/Users/Dell/Downloads/test"))