Other feed intake monitor of pig performance test station

Other feed intake monitor of pig performance test station

other_monitor(data, house_width, save_path)

Arguments

data

A data frame or data table. This is the data to be processed. It should have specific columns depending on the station type.

house_width

A character or numeric value specifying the width of the house number.If numeric, it will be converted to a character string.

save_path

A character string specifying the path where the output PNG files will be saved.

Value

A list of data.tables, each representing different aspects of station monitoring results. This list is also saved to an Excel file at the specified save_path.

Examples

# Load CSV data
data <- data.table::fread("C:/Users/Dell/Documents/projects/pptsdm_data/ppt_monitor_test_data.csv")
# Monitor station and data
res <- other_monitor(data = data, house_width = "1", save_path = "C:/Users/Dell/Downloads/test")
#> • There are no duplicate responders in different locations.
#> 
# Monitor the number of times 'na' appears in the last 7 days
res$responder_na
#>    location 24-04-07 24-04-08 24-04-09 24-04-10 24-04-11 sum_na
#>      <char>    <int>    <int>    <int>    <int>    <int>  <num>
#> 1:      502        2        1        1        3        2      9
#> 2:      503       NA        2        2       NA       NA      4
#> 3:      504        2       NA        2       NA       NA      4
#> 4:      507        1        1        1       NA       NA      3
#> 5:      508       NA       NA       NA       NA        1      1
#> 6:      509       NA       NA        1       NA        2      3
#> 7:      510       NA        1        2        1       NA      4
# Monitor the percentage of extreme weight records in the last 7 days
res$extreme_weight
#> Key: <location>
#>     location 24-04-07 24-04-08 24-04-09 24-04-10 24-04-11
#>        <int>    <num>    <num>    <num>    <num>    <num>
#>  1:      501       NA     0.86     1.01     0.93     2.48
#>  2:      502     2.11     0.97     2.06     1.28     2.38
#>  3:      503     2.41     2.08     3.61     2.60     4.17
#>  4:      504     2.33       NA       NA       NA     1.18
#>  5:      505     1.28       NA       NA     0.93     5.04
#>  6:      506     6.15     2.99       NA       NA     1.25
#>  7:      507       NA     2.53       NA       NA     3.61
#>  8:      508       NA       NA     1.32       NA       NA
#>  9:      509     1.71     0.76     0.85     0.85     3.10
#> 10:      510     1.10     1.03       NA       NA     8.33
# Monitor the visiting time and frequency of pigs in the last 7 days
res$feed_time_n
#> Key: <location>
#>     location sum_time_24-04-07 sum_time_24-04-08 sum_time_24-04-09
#>        <int>             <num>             <num>             <num>
#>  1:      501             18.09             18.14             18.81
#>  2:      502             17.28             16.42             16.92
#>  3:      503             18.15             16.93             15.49
#>  4:      504             18.65             17.66             16.92
#>  5:      505             16.82             17.51             16.73
#>  6:      506             18.65             17.29             17.48
#>  7:      507             17.45             15.67             15.85
#>  8:      508             18.12             17.93             17.49
#>  9:      509             17.97             16.96             17.39
#> 10:      510             19.18             19.84             17.95
#>     sum_time_24-04-10 sum_time_24-04-11 sum_n_24-04-07 sum_n_24-04-08
#>                 <num>             <num>          <int>          <int>
#>  1:             18.62             18.54            105            116
#>  2:             16.53             16.30             95            103
#>  3:             15.84             15.86             83             96
#>  4:             17.10             16.49             86             81
#>  5:             17.26             16.71             78             89
#>  6:             19.43             18.44             65             67
#>  7:             15.88             16.28             80             79
#>  8:             18.09             17.18             68             73
#>  9:             17.98             17.23            117            132
#> 10:             18.17             18.12             91             97
#>     sum_n_24-04-09 sum_n_24-04-10 sum_n_24-04-11
#>              <int>          <int>          <int>
#>  1:             99            108            121
#>  2:             97             78             84
#>  3:             83             77             96
#>  4:             85             74             85
#>  5:             99            107            119
#>  6:             69             78             80
#>  7:             71             83             83
#>  8:             76             76             83
#>  9:            117            118            129
#> 10:             68             69             72
# Monitor the total feed intake over the last 7 days
res$all_feedintake
#> Key: <location>
#>     location 24-04-07 24-04-08 24-04-09 24-04-10 24-04-11 sum_feed_intake
#>       <char>    <num>    <num>    <num>    <num>    <num>           <num>
#>  1:      501   41.296   43.521   46.971   45.561   46.342          223.69
#>  2:      502   46.872   46.084   49.669   46.322   50.073          239.02
#>  3:      503   51.224   49.190   47.470   48.056   48.474          244.41
#>  4:      504   48.653   49.708   48.463   48.848   47.979          243.65
#>  5:      505   43.637   46.987   47.731   47.403   44.472          230.23
#>  6:      506   47.598   46.102   45.979   50.431   47.723          237.83
#>  7:      507   44.857   41.029   43.967   43.243   44.581          217.68
#>  8:      508   44.223   45.833   43.710   44.744   43.985          222.50
#>  9:      509   48.295   48.087   51.426   50.911   49.922          248.64
#> 10:      510   43.859   47.871   44.408   44.705   46.009          226.85
# Monitor the average feed intake over the last 7 days
res$mean_feedintake
#> Key: <date>
#>        date  5_feed sum_feed_intake   5_n 5_mean_feed
#>      <char>   <num>           <num> <int>       <num>
#> 1: 24-04-07 460.514          460.51   147       3.133
#> 2: 24-04-08 464.412          464.41   146       3.181
#> 3: 24-04-09 469.794          469.79   145       3.240
#> 4: 24-04-10 470.224          470.22   146       3.221
#> 5: 24-04-11 469.560          469.56   146       3.216
# Monitor the average weight per pen over the last 7 days
res$house_weight
#> Key: <date>
#>        date house_5
#>      <char>   <num>
#> 1: 24-04-07   74.86
#> 2: 24-04-08   76.19
#> 3: 24-04-09   77.65
#> 4: 24-04-10   79.29
#> 5: 24-04-11   80.55