Feed intake monitor of pig performance test station

Feed intake monitor of pig performance test station

monitor_schedule(
  taskname,
  schedule,
  starttime,
  startdate,
  rscript_args = NULL,
  ...
)

Arguments

taskname

A character string with the name of the task. Defaults to the filename. Should not contain any spaces

schedule

Either one of 'ONCE', 'MONTHLY', 'WEEKLY', 'DAILY', 'HOURLY', 'MINUTE', 'ONLOGON', 'ONIDLE

starttime

A timepoint in HH:mm format indicating when to run the script. Defaults to within 62 seconds

startdate

A date that specifies the first date on which to run the task. Only applicable if schedule is of type 'MONTHLY', 'WEEKLY', 'DAILY', 'HOURLY', 'MINUTE'. Defaults to today in '%d/%m/%Y' format. Change to your locale format if needed

rscript_args

Character string with further arguments passed on to Rscript

...

other parameters

Value

pdf, pngs and excels

Examples

# Set monitor task
monitor_schedule(
  taskname = "ppt_csv_monitor",
  schedule = "DAILY",
  starttime = "10:05",
  startdate = format(Sys.Date(), "%Y/%m/%d"),
  rscript_args = list(house_width = "1", 
                      days = 5,
                      begin_date = "2024-05-01", 
                      csv_path = "C:/Users/Dell/Documents/projects/pptsdm_data",
                      save_path = "C:/Users/Dell/Downloads/test"))
#> [1] "SUCCESS: The scheduled task \"ppt_csv_monitor\" has successfully been created."
# Delete monitor task
taskscheduleR::taskscheduler_delete("ppt_csv_monitor")
#> [1] 0